+++ packages/api/src/ontology/ActionDefinition.ts
@@ -15,6 +15,7 @@
  */
 
 import type { OsdkMetadata } from "../OsdkMetadata.js";
+import type { InterfaceDefinition } from "./InterfaceDefinition.js";
 import type {
   ObjectTypeDefinition,
   ReleaseStatus,
@@ -43,7 +44,9 @@ export namespace ActionMetadata {
     type:
       | ValidBaseActionParameterTypes
       | DataType.Object<any>
-      | DataType.ObjectSet<any>;
+      | DataType.ObjectSet<any>
+      | DataType.Interface<any>
+      | DataType.Struct<any>;
     description?: string;
     multiplicity?: boolean;
     nullable?: boolean;
@@ -58,12 +61,25 @@ export namespace ActionMetadata {
       object: T_Target["apiName"];
     }
 
+    export interface Interface<T_Target extends InterfaceDefinition = never> {
+      __OsdkTargetType?: T_Target;
+      type: "interface";
+      interface: T_Target["apiName"];
+    }
+
     export interface ObjectSet<
       T_Target extends ObjectTypeDefinition = never,
     > {
       __OsdkTargetType?: T_Target;
       type: "objectSet";
       objectSet: T_Target["apiName"];
+    }
+
+    export interface Struct<
+      T extends Record<string, ValidBaseActionParameterTypes>,
+    > {
+      type: "struct";
+      struct: T;
     }
   }
 }
@@ -92,4 +108,5 @@ export type ValidBaseActionParameterTypes =
   | "datetime"
   | "timestamp"
   | "attachment"
-  | "marking";
+  | "marking"
+  | "objectType";