+++ packages/generator/src/v2.0/generateClientSdkVersionTwoPointZero.test.ts
@@ -487,37 +487,40 @@ describe("generator", () => {
     helper = createMockMinimalFiles();
   });
 
-  test("should be able to generate a project", async () => {
-    await generateClientSdkVersionTwoPointZero(
-      TodoWireOntology,
-      "typescript-sdk/0.0.0 osdk-cli/0.0.0",
-      helper.minimalFiles,
-      BASE_PATH,
-    );
-
-    const files = helper.getFiles();
-
-    expect(files).toMatchObject({
-      [`${BASE_PATH}/index.ts`]: expect.anything(),
-      [`${BASE_PATH}/OntologyMetadata.ts`]: expect.anything(),
-      [`${BASE_PATH}/ontology/objects/Todo.ts`]: expect.anything(),
-    });
-
-    const diagnostics = compileThis(helper.getFiles(), BASE_PATH);
-    for (const q of diagnostics) {
-      console.error(
-        `${q.file?.fileName}:${q.file?.getLineStarts()}`,
-        q.messageText,
+  test(
+    "should be able to generate a project",
+    { timeout: 20_000 },
+    async () => {
+      await generateClientSdkVersionTwoPointZero(
+        TodoWireOntology,
+        "typescript-sdk/0.0.0 osdk-cli/0.0.0",
+        helper.minimalFiles,
+        BASE_PATH,
       );
-    }
 
-    // TODO: Certain errors are expected since we can't resolve the static code, but we should fix them.
-    const errors = diagnostics.filter(q => q.code !== 2792);
-    expect(errors).toHaveLength(0);
+      const files = helper.getFiles();
 
-    expect(
-      tweakedFilesForSnapshotConsistency(helper.getFiles()),
-    ).toMatchInlineSnapshot(`
+      expect(files).toMatchObject({
+        [`${BASE_PATH}/index.ts`]: expect.anything(),
+        [`${BASE_PATH}/OntologyMetadata.ts`]: expect.anything(),
+        [`${BASE_PATH}/ontology/objects/Todo.ts`]: expect.anything(),
+      });
+
+      const diagnostics = compileThis(helper.getFiles(), BASE_PATH);
+      for (const q of diagnostics) {
+        console.error(
+          `${q.file?.fileName}:${q.file?.getLineStarts()}`,
+          q.messageText,
+        );
+      }
+
+      // TODO: Certain errors are expected since we can't resolve the static code, but we should fix them.
+      const errors = diagnostics.filter(q => q.code !== 2792);
+      expect(errors).toHaveLength(0);
+
+      expect(
+        tweakedFilesForSnapshotConsistency(helper.getFiles()),
+      ).toMatchInlineSnapshot(`
       {
         "/foo/OntologyMetadata.ts": "export type $ExpectedClientVersion = 'PLACEHOLDER';
       export const $osdkMetadata = { extraUserAgent: 'typescript-sdk/0.0.0 osdk-cli/0.0.0' };
@@ -532,6 +535,7 @@ describe("generator", () => {
       export * as $Objects from './ontology/objects';
       export { getCount, returnsTodo } from './ontology/queries';
       export * as $Queries from './ontology/queries';
+      export { $osdkMetadata } from './OntologyMetadata';
       export { $ontologyRid } from './OntologyMetadata';
       ",
         "/foo/ontology/actions.ts": "export { deleteTodos } from './actions/deleteTodos';
@@ -742,6 +746,7 @@ describe("generator", () => {
           apiName: 'SomeInterface';
           description: 'Some interface';
           displayName: 'Sum Interface';
+          implementedBy: [];
           implements: [];
           links: {};
           properties: {
@@ -1077,7 +1082,8 @@ describe("generator", () => {
       ",
       }
     `);
-  });
+    },
+  );
 
   test("throws an error when target destination is not empty", async () => {
     helper.minimalFiles.readdir = vi.fn(async (_path: string) => ["file"]);
@@ -1150,6 +1156,7 @@ describe("generator", () => {
         export * as $Objects from './ontology/objects.js';
         export { getCount, returnsTodo } from './ontology/queries.js';
         export * as $Queries from './ontology/queries.js';
+        export { $osdkMetadata } from './OntologyMetadata.js';
         export { $ontologyRid } from './OntologyMetadata.js';
         ",
           "/foo/ontology/actions.ts": "export { deleteTodos } from './actions/deleteTodos.js';
@@ -1360,6 +1367,7 @@ describe("generator", () => {
             apiName: 'foo.bar.SomeInterface';
             description: 'Some interface';
             displayName: 'Sum Interface';
+            implementedBy: [];
             implements: [];
             links: {};
             properties: {
@@ -2100,6 +2108,7 @@ describe("generator", () => {
         export * as $Objects from './ontology/objects.js';
         export {} from './ontology/queries.js';
         export * as $Queries from './ontology/queries.js';
+        export { $osdkMetadata } from './OntologyMetadata.js';
         export { $ontologyRid } from './OntologyMetadata.js';
         ",
           "/foo/ontology/actions.ts": "export {};
@@ -2151,6 +2160,7 @@ describe("generator", () => {
             strictProps: SomeInterface.StrictProps;
             apiName: 'com.example.dep.SomeInterface';
             displayName: 'Sum Interface';
+            implementedBy: [];
             implements: [];
             links: {};
             properties: {