+++ packages/shared.test/src/stubs/objects.ts
@@ -233,6 +233,7 @@ export const objectWithAllPropertyTypes1: OntologyObjectV2 = {
       type: "Polygon",
     },
   ],
+  mediaReference: "MediaReferencePlaceholder",
 };
 
 export const objectWithAllPropertyTypesEmptyEntries = {