+++ benchmarks/tests/primary/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @osdk/benchmarks.primary
 
+## 0.1.0-beta.8
+
+### Patch Changes
+
+- Updated dependencies [c80ca45]
+  - @osdk/client@2.1.0-beta.25
+
+## 0.1.0-beta.7
+
+### Patch Changes
+
+- Updated dependencies [03f95eb]
+- Updated dependencies [c691c8b]
+- Updated dependencies [3c9ac22]
+  - @osdk/client@2.1.0-beta.24
+
+## 0.1.0-beta.6
+
+### Patch Changes
+
+- Updated dependencies [d794c19]
+- Updated dependencies [7a95618]
+- Updated dependencies [62c1d63]
+- Updated dependencies [3fc2ee6]
+- Updated dependencies [d665147]
+  - @osdk/client@2.1.0-beta.23
+
 ## 0.1.0-beta.5
 
 ### Patch Changes