+++ packages/client.test.ontology/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @osdk/client.test.ontology
 
+## 2.1.0-beta.25
+
+### Patch Changes
+
+- Updated dependencies [c80ca45]
+  - @osdk/api@2.1.0-beta.25
+
+## 2.1.0-beta.24
+
+### Patch Changes
+
+- Updated dependencies [03f95eb]
+- Updated dependencies [c691c8b]
+- Updated dependencies [3c9ac22]
+  - @osdk/api@2.1.0-beta.24
+
+## 2.1.0-beta.23
+
+### Patch Changes
+
+- Updated dependencies [d794c19]
+- Updated dependencies [62c1d63]
+  - @osdk/api@2.1.0-beta.23
+
 ## 2.1.0-beta.22
 
 ### Patch Changes