+++ packages/example-generator/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @osdk/example-generator
 
+## 0.9.0-beta.24
+
+### Patch Changes
+
+- @osdk/create-app@2.1.0-beta.25
+- @osdk/create-widget@2.0.0-beta.10
+
+## 0.9.0-beta.23
+
+### Patch Changes
+
+- @osdk/create-app@2.1.0-beta.24
+- @osdk/create-widget@2.0.0-beta.9
+
+## 0.9.0-beta.22
+
+### Minor Changes
+
+- 83e7ea3: Move from a single-widget to widget set model for custom widget libraries
+
+### Patch Changes
+
+- Updated dependencies [83e7ea3]
+  - @osdk/create-widget@2.0.0-beta.8
+  - @osdk/create-app@2.1.0-beta.23
+
 ## 0.9.0-beta.21
 
 ### Patch Changes