+++ packages/generator/CHANGELOG.md
@@ -1,5 +1,46 @@
 # @osdk/generator
 
+## 2.1.0-beta.25
+
+### Patch Changes
+
+- Updated dependencies [c80ca45]
+  - @osdk/generator-converters@2.1.0-beta.25
+  - @osdk/api@2.1.0-beta.25
+
+## 2.1.0-beta.24
+
+### Minor Changes
+
+- 03f95eb: Add support for interface params in actions.
+- c691c8b: Package now has experimental support for commonjs
+
+### Patch Changes
+
+- Updated dependencies [03f95eb]
+- Updated dependencies [c691c8b]
+- Updated dependencies [3c9ac22]
+  - @osdk/generator-converters@2.1.0-beta.24
+  - @osdk/api@2.1.0-beta.24
+
+## 2.1.0-beta.23
+
+### Minor Changes
+
+- d794c19: Adding support for struct action params.
+- f0b519d: Upgrade to vitest 3 and vite 6
+- 696c59c: Exposes $osdkMetadata from the root of generated packages for clients that want access to $osdkMetadata.extraUserAgent
+- d665147: Update platform sdk deps
+
+### Patch Changes
+
+- Updated dependencies [d794c19]
+- Updated dependencies [f0b519d]
+- Updated dependencies [62c1d63]
+- Updated dependencies [d665147]
+  - @osdk/generator-converters@2.1.0-beta.23
+  - @osdk/api@2.1.0-beta.23
+
 ## 2.1.0-beta.22
 
 ### Minor Changes